feat(noise_amplification): add two-qubit noise amplifier
