+++ zne/noise_amplification/folding_amplifier/__init__.py
@@ -13,6 +13,15 @@
 """Noise amplification strategy via unitary/inverse repetition."""
 
 from .global_folding_amplifier import GlobalFoldingAmplifier
-from .local_folding_amplifier import LocalFoldingAmplifier
+from .local_folding_amplifier import (
+    CxAmplifier,
+    LocalFoldingAmplifier,
+    TwoQubitAmplifier,
+)
 
-__all__ = ["GlobalFoldingAmplifier", "LocalFoldingAmplifier"]
+__all__ = [
+    "GlobalFoldingAmplifier",
+    "LocalFoldingAmplifier",
+    "CxAmplifier",
+    "TwoQubitAmplifier",
+]
